Add a "project" working mode
One of the thing that prevents from using jhbuild outside of GNOME is
that you can effectively work only on the GNOME moduleset and the
defaults settings will not be friendly if you want to have several
projects to work with jhbuild.

Following a bit of the Android "repo" spirit, jhbuild could learn to
handle a working tree concept, where all the settings and state would
be maintained in a subdirectory.

So if a jhbuild/moduleset file is found while walking up from the
current working directory, jhbuild would use this as the default
moduleset and would configure itself with the config file next to it.